Create LSTM_trend_prediction_basic_model
This code is not a definitive model. 

This is an adapted version of the code in https://github.com/nayash/Stock-Price-Prediction/blob/master/stock_pred_main.py

There are some problems that will be adressed later, such as: overfitting, non-lagged variables and a not so appropiated forecast KPI for my intents (trend prediction evaluation).